reorganize code and update tests
